rest-api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==2.3.7
pytest==7.4.2
flask-restx==1.1.0
